// src/models/responses/car/GetByIdCarResponse.ts
export interface GetByIdCarResponse {
	id: number;
	kilometer: number;
	year: number;
	plate: string;
	dailyPrice: number;
	minFindeksRate: number;
	imagePath: string;
	trunkVolume: number;
	capacity: number;
	caseType: string;
	fuelType: string;
	gearType: string;
	modelName: string;
	colorName: string;
}